Move functions from board.py to tools.py.
