+++ docs/todo_2024.md
@@ -2,12 +2,12 @@
 # Todo List (2024)
 
 ## TODO
-- Create tools.py and move some function from board.py to tools.py
+- Use standard chess piece images
+- Test move generation: calculate number of possible positions after n moves, report time
 - Switch between king middle game and end game tables
 - Define winner and loser or draw
 - Consider putting SetStateFromPieceState() at the end of PlacePiece()
 - Write legal moves using chess notation
-- Use standard chess piece images
 - Color the square for the computer's latest move
 - For moves, color to and from squares yellow (shade based on light/dark square)
 - For possible moves, use blue squares (shade based on light/dark square)
@@ -32,3 +32,4 @@
 - Plot position evaluation maps for each piece
 - Fix plots: they are flipped over the horizontal axis
 - Add title and z color scale to plots
+- Create tools.py and move some function from board.py to tools.py
